temp fix for osx-aliases not being sourced... still need to fix plugin

+++ init/.initZsh.sh
@@ -3,6 +3,9 @@ set -e
 cd ${0%/*}
 cd .. && source dotfiles.conf
 
+color='\033[0;34m'
+NC='\033[0m' # No Color
+
 if [ ! -n "$ZSH" ]; then
   ZSH="${zsh_dst}"
 fi
@@ -27,8 +30,8 @@ dedupepath(){
 }
 
 makeZshrc(){
-    echo "\033[0;34mMaking ${zshrc} file...\033[0m"
-    mv "${zshrc}" ~/.zshrc.orig
+    echo "${color}Making ${zshrc} file...${NC}"
+    [ -e "${zshrc}" ] &&  mv "${zshrc}" ~/.zshrc.orig
     touch "${zshrc}"
     echo "## OS specific settings ############################" >> ${zshrc}
     # TODO: add profiles, rsync, ant, atom, autoenv, vim-interaction, virtualenv, command-not-found, dirhistory, & all the git plugins ???
@@ -39,7 +42,8 @@ makeZshrc(){
         #echo export TERM="xterm-256color" >> ${zshrc}
         module purge
     elif [[ "$OSTYPE" =~ ^(darwin)+ ]]; then
-        echo plugins=\(git git-extras brew sudo tmux tmuxinator colored-man vundle web-search extract zsh_reload copyfile copydir common-aliases ant fasd aws osx brew-cask xcode osx-aliases\) >> ${zshrc} 
+        #FIXME: osx-aliases doesnt work!
+        echo plugins=\(git git-extras brew sudo tmux tmuxinator colored-man vundle web-search extract zsh_reload copyfile copydir common-aliases ant fasd aws osx brew-cask xcode\) >> ${zshrc} 
         echo "export HOMEBREW_CASK_OPTS='--appdir=/Applications --caskroom=$(brew --prefix)/Caskroom'" >> ${zshrc}
         #echo source $(which zsh coreutils)/libexec/gnubin >> 
         # TODO: make this automatic
@@ -130,17 +134,23 @@ export GREP_OPTIONS='--color=auto';
 #src 
 EOF
 }
-
 linkZsh() {
-    echo "\033[0;34mLinking custom zsh  overrides...\033[0m"
+    echo "${color}Linking custom zsh overrides...${NC}"
     rm -rf $ZSH/custom
-    mkdir -p $ZSH/custom 
-    ln -sv "${zsh_src}"/* $ZSH/custom
-    #ln -sfv "${dotfiles}"/.zshrc ~
+    #ln -svv "${zsh_src}"/* $ZSH/custom
+    mkdir -p $ZSH/custom/themes
+    #mkdir -p $ZSH/custom/plugins
+    ln -svv "${zsh_src}"/*.zsh $ZSH/custom
+    #FIXME: remove this when osx-aliases not being sourced by zsh gets fixed...
+    if [[ "$OSTYPE" =~ ^(linux)+ ]]; then
+        rm -rf $ZSH/custom/osx-aliases.zsh 
+    fi
+    ln -sfv "${zsh_src}"/themes/* $ZSH/custom/themes
+    #ln -sfv "${zsh_src}"/plugins/* $ZSH/custom/plugins
 }
 
 if [ -d "$ZSH" ]; then
-  echo "\033[0;33mYou already have Oh My Zsh installed.\033[0m You'll need to remove $ZSH if you want to install"
+  echo "\033[0;33mYou already have Oh My Zsh installed.${NC} You'll need to remove $ZSH if you want to install"
   makeZshrc
   linkZsh
   exit
@@ -149,7 +159,7 @@ fi
 #TODO: check for zsh and install it using the appropriate method for current OS...
 # brew install zsh
 
-echo "\033[0;34mCloning Oh My Zsh...\033[0m"
+echo "${color}Cloning Oh My Zsh...${NC}"
 hash git >/dev/null 2>&1 && git clone git://github.com/robbyrussell/oh-my-zsh.git $ZSH || {
   echo "git not installed"
   exit
